# protos/proto/api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "protos/proto/api/api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "UrlShortener"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/url-shorter": {
      "post": {
        "operationId": "UrlShortener_CreateShortURL",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/urlshortenerCreateShortURLResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/urlshortenerCreateShortURLRequest"
            }
          }
        ],
        "tags": [
          "UrlShortener"
        ]
      }
    },
    "/url-shorter/{alias}": {
      "get": {
        "operationId": "UrlShortener_GetOriginalURL",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/urlshortenerGetOriginalURLResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UrlShortener"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "urlshortenerCreateShortURLRequest": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string"
        }
      }
    },
    "urlshortenerCreateShortURLResponse": {
      "type": "object",
      "properties": {
        "alias": {
          "type": "string"
        }
      }
    },
    "urlshortenerGetOriginalURLResponse": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string"
        }
      }
    }
  }
}